+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 
 - Add functionality to process and store span attachments. ([#5423](https://github.com/getsentry/relay/pull/5423), [#5363](https://github.com/getsentry/relay/pull/5363))
 
+**Internal**:
+
+- Revise trace metric and log size limits. ([#5440](https://github.com/getsentry/relay/pull/5440))
+
 ## 25.11.1
 
 **Breaking Changes**: